AI-Mission-Generator/requirements.txt
fastapi
uvicorn
sqlalchemy
dotenv
torch
llama-cpp-python
huggingface_hub